requirements.txt
pandas==1.0.5
wget==3.2
PyYAML==5.3.1
